+++ lib/src/synthesis.cc
@@ -14,73 +14,81 @@ std::map<char, char> Organism::rna_matrix = {
     {'C', 'G'}
 };
 
-std::string Organism::get_name() {
-    return this->name;
+std::string Organism::get_name() const {
+    return name_;
 }
 
-std::string Organism::get_rna() {
-    return this->rna;
+std::string Organism::get_rna() const {
+    return rna_;
 }
 
-std::string Organism::get_output() {
+std::string Organism::get_output() const {
     std::stringstream oput;
-    oput << "RNA sequence " << this->organism_number + 1 << ": " << this->rna << "\n\n";
-    if (!this->is_sequenced_) {
-        oput << "No start sequence found\n";
-        oput << std::string(20, '-') + "\n";
-        return oput.str();
-    }
-    oput << "Found a start sequence at " << this->start_position << std::endl;
-    oput << "Found end sequence " << this->end_codon << "\n\n";
-    for (int x = 0; x < this->rna_sep.size(); x++) 
-        oput << rna_sep[x] << " ";
-    oput << "\n\n";
-    oput << std::string(20, '-') + "\n";
-    return oput.str();
-}
+    oput << "RNA sequence " << organism_number_ + 1 << ": " << rna_ << "\n\n";
+    if (!is_sequenced_) {
+        if (!found_start_) {
+            oput << "No start sequence found, cannot sequence\n";
+            oput << std::string(20, '-') + "\n";
+            return oput.str();
+        }
+        else {
+            oput << "No end sequence found, cannot sequence\n";
+            oput << std::string(20, '-') + "\n";
+            return oput.str();
+        }
 
-std::vector<std::string> Organism::get_rna_sep() {
-    return this->rna_sep;
-}
+    }
+    oput << "Found start sequence at " << start_position_ << std::endl;
+    oput << "Found end sequence " << end_codon_ << "\n\n";
+    for (size_t x = 0; x < separated_rna_.size(); x++) 
+        oput << separated_rna_[x] << " ";
+    oput << "\n\n" << std::string(20, '-') + "\n\n";
+    return oput.str();
+ }
+
+std::vector<std::string> Organism::get_separated_rna() const {
+    return separated_rna_;
+} 
 
 void Organism::given_rna() {
-    this->rna = this->dna;
-}
+    rna_ = dna_;
+} 
 
 void Organism::translate() {
-    for (int i=0;  i<this->dna.length(); i++) 
-        this->rna += this->rna_matrix[this->dna[i]];
+    for (size_t i=0;  i<dna_.length(); i++) 
+        rna_ += rna_matrix[dna_[i]];
 }
 
 void Organism::sequence_protein() {
-    char start_sequence[] = "AUG";
-    this->start_position = this->rna.find(start_sequence);
-    if (this->start_position == std::string::npos) {
-        this->is_sequenced_ = 0;
-     }
-    std::string rna_tmp = this->rna.substr(this->start_position);
+    char start_sequence[] = "AUG"; 
+    
+    start_position_ = rna_.find(start_sequence);
+    if (start_position_ == std::string::npos) {
+        return;
+    } else { found_start_ = 1; }
+    std::string rna_tmp = rna_.substr(start_position_);
     int len = rna_tmp.length();
     int codon_count = len/3;
     rna_tmp = rna_tmp.substr(0, codon_count*3);
-    rna_sep.push_back("AUG");
-    for (int i = 1; i  < codon_count; i++) {
+    separated_rna_.push_back("AUG");
+     for (int i = 1; i  < codon_count; i++) {
         std::string codon = rna_tmp.substr(3*i, 3);
-        if (codon == "UAA" || codon  == "UAG" || codon == "UGA") {
-            rna_sep.push_back(codon);
-            this->end_codon = codon;
+        if  (codon == "UAA" || codon  == "UAG" || codon == "UGA") {
+            separated_rna_.push_back(codon);
+            end_codon_ = codon;
+            is_sequenced_ = 1;
             break;
         }
         else { 
-            rna_sep.push_back(codon);
+            separated_rna_.push_back(codon);
          }
     }
-    this->is_sequenced_ = 1;
 } 
 
-bool Organism::is_sequenced() { return this->is_sequenced_; }
+bool Organism::is_sequenced() { return is_sequenced_; }
 
-std::string compare_rna(std::vector<Organism> &orgs) {
-    int n = orgs.size();
+std::string compare_rna(const std::vector<Organism> &orgs) {
+    auto n = orgs.size();
     
     std::vector<bool> v(n);
     fill(v.begin(), v.end() - n + 2, true);
@@ -89,7 +97,7 @@ std::string compare_rna(std::vector<Organism> &orgs) {
 
     do { 
         std::vector<Organism> perm;
-        for (int i = 0; i < n; ++i) {
+        for (size_t i = 0; i < n; ++i) {
             if (v[i]) {
                 perm.push_back(orgs[i]);
              }
@@ -99,14 +107,14 @@ std::string compare_rna(std::vector<Organism> &orgs) {
 
     std::stringstream output;
 
-    for (int x = 0; x < permutations.size(); x++) {
+    for (size_t x = 0; x < permutations.size(); x++) {
         int pt_muts = 0;
         output << "Comparison between " << 
           permutations[x][0].get_name() << " and " <<
           permutations[x][1].get_name() << ":" << std::endl;
-        std::vector<std::string> rna1 = permutations[x][0].get_rna_sep();
-        std::vector<std::string> rna2 = permutations[x][1].get_rna_sep();
-        for (int i = 0; i < rna1.size(); i++) {
+        std::vector<std::string> rna1 = permutations[x][0].get_separated_rna();
+        std::vector<std::string> rna2 = permutations[x][1].get_separated_rna();
+        for (size_t i = 0; i < rna1.size(); i++) {
             if (rna1[i] != rna2[i]) {
                 output << "The strands differ at codon " <<
                   i + 1 << ": " << rna1[i] << 
@@ -121,12 +129,12 @@ std::string compare_rna(std::vector<Organism> &orgs) {
         output << "\tNumber of point mutations: " << pt_muts << "\n\n";
     }
     return output.str();
- } 
+  } 
 
 std::string calculate_orgs(std::vector<Organism> &orgs, bool rna_given=false) {
     std::vector<Organism> new_orgs;
     std::string output;
-    for (int x=0; x<orgs.size(); x++) {
+    for (size_t x = 0; x<orgs.size(); x++) {
         if (rna_given) 
             orgs[x].given_rna();
         else 
@@ -142,30 +150,25 @@ std::string calculate_orgs(std::vector<Organism> &orgs, bool rna_given=false) {
     return output;
  }
 
-void output_info(std::string &output, bool q, bool oput, std::string fn) {
-    if (oput) {
+void output_info(const std::string &output_string, bool quiet, bool output_file, std::string filename) {
+    if (output_file) {
         std::ofstream outfile;
-        outfile.open(fn);
-        outfile << output;
+        outfile.open(filename);
+        outfile << output_string;
         outfile.close();
     }
-    if (q) {
+    if (quiet) {
         return;
     }
-    std::cout << output;
+    std::cout << output_string;
 
 }
 
-void default_process(bool t, bool q, bool otf, std::string fn) {
-    int times;
-    int x=0;
+void default_process(bool test_mode, bool quiet, bool output_file, std::string filename) {
+    int x = 0;
+    auto times = test_mode ? 10000 : 1; 
     
-    if (t) 
-        times = 10000;
-    else
-        times = 1;
-    
-    do {
+    do { 
         std::vector<Organism> orgs;
         Organism afri("African", 0, "CTACGTTCATCTGGTCAGAACTGGTTA");
         Organism ande("Andes", 1, "TCACCTACGTCGATCTGGTCAGGACTT");
@@ -176,19 +179,19 @@ void default_process(bool t, bool q, bool otf, std::string fn) {
         orgs.push_back(anta);
 
         std::string outinfo = calculate_orgs(orgs);
-        output_info(outinfo, q, otf, fn);
+        output_info(outinfo, quiet, output_file, filename);
         x++;
-     } while (x < times);
+     } while (x < times); 
 } 
 
 void showhelpinfo(char *s) {
-  std::cout<<"Usage:   "<< s <<" [-option] [argument]"<<std::endl;
-  std::cout<<"option:  "<<"-h  show help information"<<std::endl;
-  std::cout<<"         "<<"-d  use default values, useful in testing"<<std::endl;
-  std::cout<<"         "<<"-q  quiet"<<std::endl;
-  std::cout<<"         "<<"-o  output file"<<std::endl;
-  std::cout<<"         "<<"-r  input is rna"<<std::endl;
-  std::cout<<"         "<<"-t  enable test mode"<<std::endl;
-  std::cout<<"         "<<"-v  show version infomation"<<std::endl;
-  std::cout<<"example: "<< s <<" -o output.txt <DNA1> <DNA2> ..."<<std::endl;
+  std::cout<<"Usage:   "<< s <<" [-option] [argument]\n"
+           <<"option:  "<<"-h  show help information\n"
+           <<"         "<<"-d  use default values, useful in testing\n"
+           <<"         "<<"-q  quiet\n"
+           <<"         "<<"-o  output file\n"
+           <<"         "<<"-r  input is rna\n"
+           <<"         "<<"-t  enable test mode\n"
+           <<"         "<<"-v  show version infomation\n"
+           <<"example: "<< s <<" -o output.txt <DNA1> <DNA2> ...\n"<<std::flush;
 }